pyproject.toml
[project]
name = "mydex-deep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "deepagents>=0.4.11",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "hyperliquid-python-sdk>=0.22.0",
    "langchain-google>=0.1.1",
    "langchain-openai>=1.1.11",
    "pydantic>=2.12.5",
    "rich>=14.3.3",
    "sse-starlette>=2.1.0",
    "uvicorn>=0.30.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
python_files = ["test_*.py", "*.test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
